Small improvements on data types and added comments; made generateRay const method

+++ output/ppm/PPMImage.cpp
@@ -6,12 +6,12 @@ PPMImage::PPMImage(const PPMImageMeta& metadata): metadata(metadata)
 {
   buffer.resize(metadata.imgHeight); // TODO: Test with big allocations
 
-  for (int32_t i = 0; i < metadata.imgHeight; i++) {
+  for (int16_t i = 0; i < metadata.imgHeight; i++) {
     buffer[i].resize(metadata.imgWidth);
   }
 }
 
-void PPMImage::writePixel(int32_t row, int32_t col, const PPMColor& color)
+void PPMImage::writePixel(uint16_t row, uint16_t col, const PPMColor& color)
 {
   buffer[row][col] = color;
 }
